// src/gen-ts/src/models/CrupdateLeaveType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
/**
 *
 * @export
 * @interface CrupdateLeaveType
 */
export interface CrupdateLeaveType {
  /**
   *
   * @type {string}
   * @memberof CrupdateLeaveType
   */
  id?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateLeaveType
   */
  name?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateLeaveType
   */
  description?: string
  /**
   *
   * @type {boolean}
   * @memberof CrupdateLeaveType
   */
  paid?: boolean
  /**
   *
   * @type {boolean}
   * @memberof CrupdateLeaveType
   */
  deduct_from_balance?: boolean
  /**
   *
   * @type {string}
   * @memberof CrupdateLeaveType
   */
  color?: string
  /**
   *
   * @type {number}
   * @memberof CrupdateLeaveType
   */
  days_per_year?: number
  /**
   *
   * @type {string}
   * @memberof CrupdateLeaveType
   */
  company_id?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateLeaveType
   */
  comment?: string
}

/**
 * Check if a given object implements the CrupdateLeaveType interface.
 */
export function instanceOfCrupdateLeaveType(value: object): value is CrupdateLeaveType {
  return true
}

export function CrupdateLeaveTypeFromJSON(json: any): CrupdateLeaveType {
  return CrupdateLeaveTypeFromJSONTyped(json, false)
}

export function CrupdateLeaveTypeFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): CrupdateLeaveType {
  if (json == null) {
    return json
  }
  return {
    id: json['id'] == null ? undefined : json['id'],
    name: json['name'] == null ? undefined : json['name'],
    description: json['description'] == null ? undefined : json['description'],
    paid: json['paid'] == null ? undefined : json['paid'],
    deduct_from_balance:
      json['deduct_from_balance'] == null ? undefined : json['deduct_from_balance'],
    color: json['color'] == null ? undefined : json['color'],
    days_per_year: json['days_per_year'] == null ? undefined : json['days_per_year'],
    company_id: json['company_id'] == null ? undefined : json['company_id'],
    comment: json['comment'] == null ? undefined : json['comment'],
  }
}

export function CrupdateLeaveTypeToJSON(json: any): CrupdateLeaveType {
  return CrupdateLeaveTypeToJSONTyped(json, false)
}

export function CrupdateLeaveTypeToJSONTyped(
  value?: CrupdateLeaveType | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    id: value['id'],
    name: value['name'],
    description: value['description'],
    paid: value['paid'],
    deduct_from_balance: value['deduct_from_balance'],
    color: value['color'],
    days_per_year: value['days_per_year'],
    company_id: value['company_id'],
    comment: value['comment'],
  }
}